feat: 메인, qna 페이지 분리
- 메인과 qua 페이지 도메인 분리
