+++ src/components/layout/navigation/Navigation.tsx
@@ -2,7 +2,6 @@
 
 import {
   usePathname,
-  useRouter,
   useSearchParams,
   useSelectedLayoutSegment,
 } from "next/navigation"
@@ -15,17 +14,6 @@ function Navigation() {
   const currentSegment = useSelectedLayoutSegment()
   const pathname = usePathname()
   const searchParmas = useSearchParams()
-
-  const { replace } = useRouter()
-
-  if (currentSegment === null && pathname === "/") {
-    if (!searchParmas.get("page")) {
-      const urlSearchParams = new URLSearchParams()
-      urlSearchParams.set("page", "0")
-
-      replace(`/?${urlSearchParams.toString()}`)
-    }
-  }
 
   // NotFoundPage
   if (currentSegment === NOTMATCH_SEGMENT) return null
