+++ src/constants/layoutMeta.ts
@@ -1,5 +1,6 @@
 export type LayoutMetaKey =
   | "/"
+  | "qna"
   | "question"
   | "updateQuestion"
   | "signup"
@@ -24,8 +25,17 @@ export const NOTMATCH_SEGMENT = "__DEFAULT__"
 
 export const layoutMeta = {
   "/": {
+    title: "홈",
+    description: "커널 스퀘어에 오신 것을 환영합니다",
+    containLayout: {
+      header: true,
+      navigation: true,
+      footer: false,
+    },
+  },
+  qna: {
     title: "개발자 Q&A",
-    description: "Q&A",
+    description: "커널 스퀘어 개발자 Q&A",
     containLayout: {
       header: true,
       navigation: true,
